+++ pool.cpp
@@ -6,7 +6,7 @@
 #include "memfile.hpp"
 #include "pool.hpp"
 
-int swizzlecmp(const char *a, const char *b) {
+int32_t swizzlecmp(const char *a, const char *b) {
 	ssize_t alen = strlen(a);
 	ssize_t blen = strlen(b);
 
@@ -14,7 +14,7 @@ int swizzlecmp(const char *a, const char *b) {
 		return 0;
 	}
 
-	long long hash1 = 0, hash2 = 0;
+	int64_t hash1 = 0, hash2 = 0;
 	for (ssize_t i = alen - 1; i >= 0; i--) {
 		hash1 = (hash1 * 37 + a[i]) & INT_MAX;
 	}
@@ -22,7 +22,7 @@ int swizzlecmp(const char *a, const char *b) {
 		hash2 = (hash2 * 37 + b[i]) & INT_MAX;
 	}
 
-	int h1 = hash1, h2 = hash2;
+	int32_t h1 = hash1, h2 = hash2;
 	if (h1 == h2) {
 		return strcmp(a, b);
 	}
@@ -30,8 +30,8 @@ int swizzlecmp(const char *a, const char *b) {
 	return h1 - h2;
 }
 
-long long addpool(struct memfile *poolfile, struct memfile *treefile, const char *s, char type) {
-	unsigned long *sp = &treefile->tree;
+int64_t addpool(struct memfile *poolfile, struct memfile *treefile, const char *s, char type) {
+	uint64_t *sp = &treefile->tree;
 	size_t depth = 0;
 
 	// In typical data, traversal depth generally stays under 2.5x
@@ -41,7 +41,7 @@ long long addpool(struct memfile *poolfile, struct memfile *treefile, const char
 	}
 
 	while (*sp != 0) {
-		int cmp = swizzlecmp(s, poolfile->map + ((struct stringpool *) (treefile->map + *sp))->off + 1);
+		int32_t cmp = swizzlecmp(s, poolfile->map + ((struct stringpool *) (treefile->map + *sp))->off + 1);
 
 		if (cmp == 0) {
 			cmp = type - (poolfile->map + ((struct stringpool *) (treefile->map + *sp))->off)[0];
@@ -60,7 +60,7 @@ long long addpool(struct memfile *poolfile, struct memfile *treefile, const char
 			// Search is very deep, so string is probably unique.
 			// Add it to the pool without adding it to the search tree.
 
-			long long off = poolfile->off;
+			int64_t off = poolfile->off;
 			if (memfile_write(poolfile, &type, 1) < 0) {
 				perror("memfile write");
 				exit(EXIT_FAILURE);
@@ -74,14 +74,14 @@ long long addpool(struct memfile *poolfile, struct memfile *treefile, const char
 	}
 
 	// *sp is probably in the memory-mapped file, and will move if the file grows.
-	long long ssp;
+	int64_t ssp;
 	if (sp == &treefile->tree) {
 		ssp = -1;
 	} else {
 		ssp = ((char *) sp) - treefile->map;
 	}
 
-	long long off = poolfile->off;
+	int64_t off = poolfile->off;
 	if (memfile_write(poolfile, &type, 1) < 0) {
 		perror("memfile write");
 		exit(EXIT_FAILURE);
@@ -106,7 +106,7 @@ long long addpool(struct memfile *poolfile, struct memfile *treefile, const char
 	tsp.right = 0;
 	tsp.off = off;
 
-	long long p = treefile->off;
+	int64_t p = treefile->off;
 	if (memfile_write(treefile, &tsp, sizeof(struct stringpool)) < 0) {
 		perror("memfile write");
 		exit(EXIT_FAILURE);
@@ -115,7 +115,7 @@ long long addpool(struct memfile *poolfile, struct memfile *treefile, const char
 	if (ssp == -1) {
 		treefile->tree = p;
 	} else {
-		*((long long *) (treefile->map + ssp)) = p;
+		*((int64_t *) (treefile->map + ssp)) = p;
 	}
 	return off;
 }
